build: automate setting of Grafana password when deploy-simple.sh is run for the first time
Using k8s' configmap/secret for the Grafana password management turned out to be complicated. Mostly because we are not using "pure" Grafana, but a third-party implementation through a Helm chart. There is only so much we can do without having to clone their repo. I've therefore opted for an automated ".dist" solution through the deploy-simple.sh bash script. The password is automatically set in values.yml through the script, which is added to gitignore to prevent accidentally committing any credentials.