build(cz): use cz output for container version
The publish-docker-image job no longer runs on tag push, but instead on
every push to main. As a consequence, `GITHUB_REF` no longer refers to
the latest tag. Additionally, the action runs checkout for the commit
that triggered the action, not the commit pushed by the `bump_version`
job. Luckily GitHub actions supports job outputs, and the commitizen
action gives the new version as such an output. Use that version as the
version of the image to be pushed.